+++ src/pages/[network]/execute.tsx
@@ -1,3 +1,0 @@
-import Execute from "lib/pages/execute";
-
-export default Execute;